refactor(program): modularize lib.rs into state/errors/instructions modules, clean TODO comments from tests
