+++ programs/stablecoin/src/instructions/burn.rs
@@ -1,0 +1,19 @@
+use crate::state::StablecoinConfig;
+use anchor_lang::prelude::*;
+use anchor_spl::token_interface::{Mint, TokenAccount};
+
+#[derive(Accounts)]
+pub struct BurnTokens<'info> {
+    #[account(
+        mut,
+        seeds = [b"config", mint.key().as_ref()],
+        bump = config.bump
+    )]
+    pub config: Account<'info, StablecoinConfig>,
+    #[account(mut)]
+    pub mint: InterfaceAccount<'info, Mint>,
+    #[account(mut, token::mint = mint)]
+    pub source_token: InterfaceAccount<'info, TokenAccount>,
+    pub authority: Signer<'info>,
+    pub token_program: Program<'info, anchor_spl::token_2022::Token2022>,
+}
