+++ programs/stablecoin/src/instructions/freeze_thaw.rs
@@ -1,0 +1,33 @@
+use crate::state::StablecoinConfig;
+use anchor_lang::prelude::*;
+use anchor_spl::token_interface::{Mint, TokenAccount};
+
+#[derive(Accounts)]
+pub struct FreezeAccount<'info> {
+    #[account(
+        mut,
+        seeds = [b"config", mint.key().as_ref()],
+        bump = config.bump
+    )]
+    pub config: Account<'info, StablecoinConfig>,
+    pub mint: InterfaceAccount<'info, Mint>,
+    #[account(mut, token::mint = mint)]
+    pub token_account: InterfaceAccount<'info, TokenAccount>,
+    pub signer: Signer<'info>,
+    pub token_program: Program<'info, anchor_spl::token_2022::Token2022>,
+}
+
+#[derive(Accounts)]
+pub struct ThawAccount<'info> {
+    #[account(
+        mut,
+        seeds = [b"config", mint.key().as_ref()],
+        bump = config.bump
+    )]
+    pub config: Account<'info, StablecoinConfig>,
+    pub mint: InterfaceAccount<'info, Mint>,
+    #[account(mut, token::mint = mint)]
+    pub token_account: InterfaceAccount<'info, TokenAccount>,
+    pub signer: Signer<'info>,
+    pub token_program: Program<'info, anchor_spl::token_2022::Token2022>,
+}
