+++ programs/stablecoin/src/instructions/mint.rs
@@ -1,0 +1,30 @@
+use crate::state::{MinterAllowance, StablecoinConfig};
+use anchor_lang::prelude::*;
+use anchor_spl::token_interface::{Mint, TokenAccount};
+
+#[derive(Accounts)]
+pub struct MintTokens<'info> {
+    #[account(
+        mut,
+        seeds = [b"config", mint.key().as_ref()],
+        bump = config.bump
+    )]
+    pub config: Account<'info, StablecoinConfig>,
+    #[account(mut)]
+    pub mint: InterfaceAccount<'info, Mint>,
+    #[account(
+        init_if_needed,
+        seeds = [b"minter", mint.key().as_ref(), minter.key().as_ref()],
+        bump,
+        space = 8 + MinterAllowance::INIT_SPACE,
+        payer = payer
+    )]
+    pub minter_pda: Account<'info, MinterAllowance>,
+    pub minter: Signer<'info>,
+    #[account(mut, token::mint = mint)]
+    pub destination_token: InterfaceAccount<'info, TokenAccount>,
+    #[account(mut)]
+    pub payer: Signer<'info>,
+    pub system_program: Program<'info, System>,
+    pub token_program: Program<'info, anchor_spl::token_2022::Token2022>,
+}
