+++ programs/stablecoin/src/instructions/mod.rs
@@ -1,0 +1,15 @@
+pub mod burn;
+pub mod compliance;
+pub mod freeze_thaw;
+pub mod initialize;
+pub mod mint;
+pub mod pause;
+pub mod roles;
+
+pub use burn::BurnTokens;
+pub use compliance::{AddToBlacklist, InitializeExtraAccountMetaList, RemoveFromBlacklist, Seize};
+pub use freeze_thaw::{FreezeAccount, ThawAccount};
+pub use initialize::Initialize;
+pub use mint::MintTokens;
+pub use pause::{Pause, Unpause};
+pub use roles::{AcceptAuthority, SetMinterAllowance, UpdateRoles};
