+++ programs/stablecoin/src/instructions/pause.rs
@@ -1,0 +1,27 @@
+use crate::state::StablecoinConfig;
+use anchor_lang::prelude::*;
+use anchor_spl::token_interface::Mint;
+
+#[derive(Accounts)]
+pub struct Pause<'info> {
+    #[account(
+        mut,
+        seeds = [b"config", mint.key().as_ref()],
+        bump = config.bump
+    )]
+    pub config: Account<'info, StablecoinConfig>,
+    pub mint: InterfaceAccount<'info, Mint>,
+    pub signer: Signer<'info>,
+}
+
+#[derive(Accounts)]
+pub struct Unpause<'info> {
+    #[account(
+        mut,
+        seeds = [b"config", mint.key().as_ref()],
+        bump = config.bump
+    )]
+    pub config: Account<'info, StablecoinConfig>,
+    pub mint: InterfaceAccount<'info, Mint>,
+    pub signer: Signer<'info>,
+}
