+++ programs/stablecoin/src/lib.rs
@@ -1,88 +1,22 @@
 #![allow(unused_variables)]
+
+mod errors;
+mod instructions;
+mod state;
+
 use anchor_lang::prelude::*;
-use anchor_spl::{
-    token_2022::Token2022,
-    token_interface::{Mint, TokenAccount},
+use anchor_spl::token_2022::Token2022;
+use anchor_spl::token_interface::{Mint, TokenAccount};
+
+use errors::StablecoinError;
+use instructions::{
+    AcceptAuthority, AddToBlacklist, BurnTokens, FreezeAccount, Initialize,
+    InitializeExtraAccountMetaList, MintTokens, Pause, RemoveFromBlacklist, Seize,
+    SetMinterAllowance, ThawAccount, Unpause, UpdateRoles,
 };
+use state::{BlacklistEntry, InitializeArgs, MinterAllowance, StablecoinConfig, UpdateRolesArgs};
 
 declare_id!("2N19eMKD2xGpjNzfktVCPnkrbGJZAzuDFoH7SJtQiNm9");
-
-#[error_code]
-pub enum StablecoinError {
-    #[msg("Unauthorized: caller does not have the required authority")]
-    Unauthorized,
-    #[msg("Not Compliant Stablecoin: SSS-2 feature not enabled")]
-    NotCompliantStablecoin,
-    #[msg("Overflow Error: arithmetic overflow occurred")]
-    OverflowError,
-    #[msg("Account Not Frozen: account must be frozen before seize")]
-    AccountNotFrozen,
-    #[msg("Invalid Amount: amount must be greater than zero")]
-    InvalidAmount,
-    #[msg("Allowance Exceeded: mint allowance exceeded")]
-    AllowanceExceeded,
-    #[msg("Invalid String Length: string exceeds maximum character limit")]
-    InvalidStringLength,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone)]
-pub struct InitializeArgs {
-    pub name: String,
-    pub symbol: String,
-    pub uri: String,
-    pub master_authority: Pubkey,
-    pub master_minter: Pubkey,
-    pub blacklister: Pubkey,
-    pub pauser: Pubkey,
-    pub enable_permanent_delegate: bool,
-    pub enable_transfer_hook: bool,
-    pub default_account_frozen: bool,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone)]
-pub struct UpdateRolesArgs {
-    pub new_master_authority: Option<Pubkey>,
-    pub new_master_minter: Option<Pubkey>,
-    pub new_blacklister: Option<Pubkey>,
-    pub new_pauser: Option<Pubkey>,
-}
-
-#[account]
-#[derive(InitSpace)]
-pub struct StablecoinConfig {
-    #[max_len(32)]
-    pub name: String,
-    #[max_len(10)]
-    pub symbol: String,
-    #[max_len(200)]
-    pub uri: String,
-    pub decimals: u8,
-    pub master_authority: Pubkey,
-    pub pending_master_authority: Option<Pubkey>,
-    pub master_minter: Pubkey,
-    pub blacklister: Pubkey,
-    pub pauser: Pubkey,
-    pub is_paused: bool,
-    pub total_supply: u64,
-    pub enable_permanent_delegate: bool,
-    pub enable_transfer_hook: bool,
-    pub default_account_frozen: bool,
-    pub bump: u8,
-}
-
-#[account]
-#[derive(InitSpace)]
-pub struct MinterAllowance {
-    pub allowance: u64,
-    pub bump: u8,
-}
-
-#[account]
-#[derive(InitSpace)]
-pub struct BlacklistEntry {
-    pub is_blacklisted: bool,
-    pub bump: u8,
-}
 
 #[program]
 pub mod stablecoin {
@@ -403,261 +337,4 @@ pub mod stablecoin {
         }
         Ok(())
     }
-}
-
-#[derive(Accounts)]
-#[instruction(decimals: u8)]
-pub struct Initialize<'info> {
-    #[account(
-        init,
-        payer = payer,
-        seeds = [b"config", mint.key().as_ref()],
-        bump,
-        space = 8 + StablecoinConfig::INIT_SPACE
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    #[account(
-        init,
-        payer = payer,
-        mint::authority = config,
-        mint::freeze_authority = config,
-        mint::decimals = decimals
-    )]
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token2022>,
-}
-
-#[derive(Accounts)]
-pub struct MintTokens<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    #[account(mut)] // FIXED: mint_to CPI must write to mint to update supply
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(
-        init_if_needed,
-        seeds = [b"minter", mint.key().as_ref(), minter.key().as_ref()],
-        bump,
-        space = 8 + MinterAllowance::INIT_SPACE,
-        payer = payer
-    )]
-    pub minter_pda: Account<'info, MinterAllowance>,
-    pub minter: Signer<'info>,
-    #[account(mut, token::mint = mint)]
-    pub destination_token: InterfaceAccount<'info, TokenAccount>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token2022>,
-}
-
-#[derive(Accounts)]
-pub struct BurnTokens<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    #[account(mut)] // FIXED: burn CPI must write to mint to update supply
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(mut, token::mint = mint)]
-    pub source_token: InterfaceAccount<'info, TokenAccount>,
-    pub authority: Signer<'info>,
-    pub token_program: Program<'info, Token2022>,
-}
-
-#[derive(Accounts)]
-pub struct FreezeAccount<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(mut, token::mint = mint)]
-    pub token_account: InterfaceAccount<'info, TokenAccount>,
-    pub signer: Signer<'info>,
-    pub token_program: Program<'info, Token2022>,
-}
-
-#[derive(Accounts)]
-pub struct ThawAccount<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(mut, token::mint = mint)]
-    pub token_account: InterfaceAccount<'info, TokenAccount>,
-    pub signer: Signer<'info>,
-    pub token_program: Program<'info, Token2022>,
-}
-
-#[derive(Accounts)]
-pub struct Pause<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    pub signer: Signer<'info>,
-}
-
-#[derive(Accounts)]
-pub struct Unpause<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    pub signer: Signer<'info>,
-}
-
-#[derive(Accounts)]
-pub struct UpdateRoles<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    pub signer: Signer<'info>,
-}
-
-#[derive(Accounts)]
-pub struct AcceptAuthority<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    pub signer: Signer<'info>,
-}
-
-#[derive(Accounts)]
-pub struct AddToBlacklist<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(
-        init,
-        seeds = [b"blacklist", mint.key().as_ref(), target_address.key().as_ref()],
-        bump,
-        space = 8 + BlacklistEntry::INIT_SPACE,
-        payer = payer
-    )]
-    pub blacklist_entry: Account<'info, BlacklistEntry>,
-    /// CHECK: target address is only used as a seed for PDA derivation
-    pub target_address: UncheckedAccount<'info>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    pub signer: Signer<'info>,
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct RemoveFromBlacklist<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(
-        mut,
-        seeds = [b"blacklist", mint.key().as_ref(), target_address.key().as_ref()],
-        bump = blacklist_entry.bump
-    )]
-    pub blacklist_entry: Account<'info, BlacklistEntry>,
-    /// CHECK: target address is only used as a seed for PDA derivation
-    pub target_address: UncheckedAccount<'info>,
-    pub signer: Signer<'info>,
-}
-
-#[derive(Accounts)]
-pub struct Seize<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(mut, token::mint = mint)]
-    pub source_token: InterfaceAccount<'info, TokenAccount>,
-    #[account(mut, token::mint = mint)]
-    pub destination_token: InterfaceAccount<'info, TokenAccount>,
-    pub signer: Signer<'info>,
-    pub token_program: Program<'info, Token2022>,
-}
-
-#[derive(Accounts)]
-pub struct SetMinterAllowance<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    #[account(
-        init_if_needed,
-        seeds = [b"minter", mint.key().as_ref(), minter.key().as_ref()],
-        bump,
-        space = 8 + MinterAllowance::INIT_SPACE,
-        payer = payer
-    )]
-    pub minter_allowance: Account<'info, MinterAllowance>,
-    /// CHECK: minter address is only used as a seed for PDA derivation
-    pub minter: UncheckedAccount<'info>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    pub signer: Signer<'info>,
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct InitializeExtraAccountMetaList<'info> {
-    #[account(
-        mut,
-        seeds = [b"config", mint.key().as_ref()],
-        bump = config.bump
-    )]
-    pub config: Account<'info, StablecoinConfig>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    pub signer: Signer<'info>,
-    #[account(
-        init,
-        seeds = [b"extra-account-metas", mint.key().as_ref()],
-        bump,
-        space = 8 + 1024,
-        payer = payer
-    )]
-    /// CHECK: extra account meta list for transfer hook
-    pub extra_account_meta_list: AccountInfo<'info>,
-    pub mint: InterfaceAccount<'info, Mint>,
-    pub system_program: Program<'info, System>,
 }
