+++ tests/sss-2.test.ts
@@ -28,10 +28,8 @@ describe("SSS-2: Compliant Stablecoin", () => {
   const treasury = Keypair.generate().publicKey;
 
   before(async () => {
-    // TODO: Generate a new mint keypair for the SSS-2 test
     mint = Keypair.generate();
     
-    // TODO: Derive the config PDA using the mint address
     [configPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("config"), mint.publicKey.toBuffer()],
       STABLECOIN_PROGRAM_ID
@@ -39,10 +37,8 @@ describe("SSS-2: Compliant Stablecoin", () => {
   });
 
   it("initialize → verify config has enablePermanentDelegate=true", async () => {
-    // TODO: Get SSS-2 preset configuration
     const preset = Presets.SSS_2;
     
-    // TODO: Initialize the stablecoin with SSS-2 preset (compliance enabled)
     const decimals = 6;
     const name = "Compliant SSS-2";
     const symbol = "CSTB2";
@@ -73,18 +69,14 @@ describe("SSS-2: Compliant Stablecoin", () => {
 
     console.log("Initialize SSS-2 tx:", tx);
 
-    // TODO: Fetch the config account and verify flags
     const config = await program.account.stablecoinConfig.fetch(configPDA);
     
-    // TODO: Assert enablePermanentDelegate === true
     assert.equal(config.enablePermanentDelegate, true, "SSS-2 should have enablePermanentDelegate=true");
     
-    // TODO: Assert enableTransferHook === true
     assert.equal(config.enableTransferHook, true, "SSS-2 should have enableTransferHook=true");
   });
 
   it("setMinterAllowance → mint 1,000,000 tokens", async () => {
-    // TODO: Set minter allowance to 2_000_000 for the master authority
     const allowance = 2_000_000;
     const [minterPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("minter"), mint.publicKey.toBuffer(), masterAuthority.toBuffer()],
@@ -104,7 +96,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
       })
       .rpc();
 
-    // TODO: Create user token account
     userTokenAccount = await getAssociatedTokenAddress(
       mint.publicKey,
       user,
@@ -121,7 +112,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
     );
     await program.provider.sendAndConfirm!(new anchor.web3.Transaction().add(userATAInstr));
 
-    // TODO: Mint 1_000_000 tokens to the user
     const mintAmount = 1_000_000;
     await program.methods
       .mint(new anchor.BN(mintAmount))
@@ -137,20 +127,17 @@ describe("SSS-2: Compliant Stablecoin", () => {
       })
       .rpc();
 
-    // TODO: Verify balance
     const accountInfo = await program.provider.connection.getParsedAccountInfo(userTokenAccount);
     const balance = (accountInfo.value?.data as any)?.parsed?.info?.tokenAmount?.uiAmount || 0;
     assert.equal(balance, 1_000_000, "User should have 1,000,000 tokens");
   });
 
   it("addToBlacklist → verify BlacklistEntry PDA created", async () => {
-    // TODO: Derive the blacklist entry PDA for blacklistedUser
     const [blacklistPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("blacklist"), mint.publicKey.toBuffer(), blacklistedUser.toBuffer()],
       STABLECOIN_PROGRAM_ID
     );
 
-    // TODO: Create token account for blacklisted user
     blacklistedTokenAccount = await getAssociatedTokenAddress(
       mint.publicKey,
       blacklistedUser,
@@ -167,7 +154,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
     );
     await program.provider.sendAndConfirm!(new anchor.web3.Transaction().add(blacklistedATAInstr));
 
-    // TODO: Mint some tokens to blacklisted user first
     const [minterPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("minter"), mint.publicKey.toBuffer(), masterAuthority.toBuffer()],
       STABLECOIN_PROGRAM_ID
@@ -187,7 +173,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
       })
       .rpc();
 
-    // TODO: Call addToBlacklist instruction
     const tx = await program.methods
       .addToBlacklist()
       .accounts({
@@ -203,13 +188,11 @@ describe("SSS-2: Compliant Stablecoin", () => {
 
     console.log("Add to blacklist tx:", tx);
 
-    // TODO: Fetch and verify BlacklistEntry PDA was created with isBlacklisted=true
     const blacklistEntry = await program.account.blacklistEntry.fetch(blacklistPDA);
     assert.equal(blacklistEntry.isBlacklisted, true, "User should be blacklisted");
   });
 
   it("transfer blocked for blacklisted address (hook fires)", async () => {
-    // TODO: Create treasury token account
     treasuryTokenAccount = await getAssociatedTokenAddress(
       mint.publicKey,
       treasury,
@@ -226,7 +209,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
     );
     await program.provider.sendAndConfirm!(new anchor.web3.Transaction().add(treasuryATAInstr));
 
-    // TODO: Attempt transfer from blacklisted address
     try {
       const transferIx = createTransferCheckedInstruction(
         blacklistedTokenAccount,
@@ -244,7 +226,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
       );
       assert.fail("Transfer should have been blocked for blacklisted address");
     } catch (e: any) {
-      // TODO: Expect error containing 'BlacklistedAddress'
       const errorMsg = e.toString().toLowerCase();
       assert.isTrue(
         errorMsg.includes("blacklist") || errorMsg.includes("blocked") || errorMsg.includes("denied"),
@@ -255,7 +236,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
   });
 
   it("freeze → seize (requires freeze first) → verify permanent delegate CPI", async () => {
-    // TODO: Freeze the blacklisted account first
     await program.methods
       .freezeAccount()
       .accounts({
@@ -269,7 +249,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
 
     console.log("Account frozen");
 
-    // TODO: Seize tokens from the frozen account
     const seizeAmount = 100000;
     const tx = await program.methods
       .seize(new anchor.BN(seizeAmount))
@@ -285,12 +264,10 @@ describe("SSS-2: Compliant Stablecoin", () => {
 
     console.log("Seize tx:", tx);
 
-    // TODO: Verify tokens were transferred to treasury
     const treasuryInfo = await program.provider.connection.getParsedAccountInfo(treasuryTokenAccount);
     const treasuryBalance = (treasuryInfo.value?.data as any)?.parsed?.info?.tokenAmount?.uiAmount || 0;
     assert.isTrue(treasuryBalance >= 0.1, "Treasury should have seized tokens");
 
-    // TODO: Attempt seize without freeze first - should fail with AccountNotFrozen
     const blacklistedUser2 = Keypair.generate().publicKey;
     const blacklistedTokenAccount2 = await getAssociatedTokenAddress(
       mint.publicKey,
@@ -342,19 +319,16 @@ describe("SSS-2: Compliant Stablecoin", () => {
         .rpc();
       assert.fail("Seize should fail without freeze first");
     } catch (e: any) {
-      // TODO: Expect error about account not frozen
       console.log("Seize without freeze failed as expected");
     }
   });
 
   it("removeFromBlacklist → PDA closed, lamports refunded", async () => {
-    // TODO: Get the blacklist entry PDA
     const [blacklistPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("blacklist"), mint.publicKey.toBuffer(), blacklistedUser.toBuffer()],
       STABLECOIN_PROGRAM_ID
     );
 
-    // TODO: Call removeFromBlacklist instruction
     const tx = await program.methods
       .removeFromBlacklist()
       .accounts({
@@ -368,20 +342,17 @@ describe("SSS-2: Compliant Stablecoin", () => {
 
     console.log("Remove from blacklist tx:", tx);
 
-    // TODO: Verify the blacklist entry is now marked as not blacklisted
     const blacklistEntry = await program.account.blacklistEntry.fetch(blacklistPDA);
     assert.equal(blacklistEntry.isBlacklisted, false, "User should be removed from blacklist");
   });
 
   it("SSS-2 gate: addToBlacklist fails on SSS-1 mint with NotCompliantStablecoin", async () => {
-    // TODO: Create a new SSS-1 mint for this test
     const sss1Mint = Keypair.generate();
     const [sss1ConfigPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("config"), sss1Mint.publicKey.toBuffer()],
       STABLECOIN_PROGRAM_ID
     );
 
-    // TODO: Initialize as SSS-1 (enablePermanentDelegate: false)
     await program.methods
       .initialize(6, {
         name: "SSS-1 Coin",
@@ -405,7 +376,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
       .signers([sss1Mint])
       .rpc();
 
-    // TODO: Attempt to call addToBlacklist on SSS-1 mint
     const [sss1BlacklistPDA] = PublicKey.findProgramAddressSync(
       [Buffer.from("blacklist"), sss1Mint.publicKey.toBuffer(), blacklistedUser.toBuffer()],
       STABLECOIN_PROGRAM_ID
@@ -426,7 +396,6 @@ describe("SSS-2: Compliant Stablecoin", () => {
         .rpc();
       assert.fail("addToBlacklist should fail on SSS-1 mint");
     } catch (e: any) {
-      // TODO: Expect error containing 'NotCompliantStablecoin' or similar
       const errorMsg = e.toString().toLowerCase();
       assert.isTrue(
         errorMsg.includes("notcompliantstablecoin") || errorMsg.includes("sss-2") || errorMsg.includes("compliance"),
